Update patch set 2
Patch Set 2:

If you can confirm it behaves well on 2.6 as-is, we can consider a stable-2.6, stable-2.7, etc branch.  I am working on 2.7, so I have not taken the time to verify that the code works elsewhere.  I think it should be a pom.xml change, as long as the API I am using is functionally equivalent between branches.  I'm undecided on what branching & tagging schema would work best for this plugin, as well - suggestions welcome!

Patch-set: 2