Update patch set 1
Patch Set 1:

Healthcheck right now is mostly a blank slate; I started adding some things to it because the original purpose for it became obsolete, but I found some basic consistency checks we could automate in this manner, so I started going down that road.

I don't have a problem considering this for admin-console (even if I'm not sure it fits in well with admin-console's current goals), but I'm more concerned with the stability of such actions on a live instance.  At the very least, sanity checks to ensure multiple reindexers are not running simultaneously are mandatory.

Patch-set: 1